hab: Refactor common exporter code to reduce LOC
All the exporter modules were simply duplicating most of the code and it
was starting to become a mess so let's put the common code in one module
before adding another exporter (Helm) module.
